[src] Create directory for optimizer
